requirements.txt
jupyter
torch
torchvision
torchmetrics
numpy
pandas
matplotlib
scikit-learn
optuna
pillow
transformers
einops
pyswarm

